+++ tests/test.rs
@@ -3,44 +3,25 @@
 mod auxiliary;
 
 use anyhow::Context as _;
-use auxiliary::test_report;
+use auxiliary::{cargo_llvm_cov, test_report, CommandExt};
+use fs_err as fs;
+
+fn test_set() -> Vec<(&'static str, &'static [&'static str])> {
+    vec![
+        ("txt", &["--text"]),
+        ("hide-instantiations.txt", &["--text", "--hide-instantiations"]),
+        ("summary.txt", &[]),
+        ("json", &["--json", "--summary-only"]),
+        ("full.json", &["--json"]),
+        ("lcov.info", &["--lcov", "--summary-only"]),
+    ]
+}
 
 fn run(model: &str, name: &str, args: &[&str]) {
     let id = format!("{}/{}", model, name);
-
-    test_report(model, name, "txt", {
-        let mut v = vec!["--text"];
-        v.extend_from_slice(args);
-        v
-    })
-    .context(id.clone())
-    .unwrap();
-
-    test_report(model, name, "summary.txt", args).context(id.clone()).unwrap();
-
-    test_report(model, name, "json", {
-        let mut v = vec!["--json", "--summary-only"];
-        v.extend_from_slice(args);
-        v
-    })
-    .context(id.clone())
-    .unwrap();
-
-    test_report(model, name, "full.json", {
-        let mut v = vec!["--json"];
-        v.extend_from_slice(args);
-        v
-    })
-    .context(id.clone())
-    .unwrap();
-
-    test_report(model, name, "lcov.info", {
-        let mut v = vec!["--lcov", "--summary-only"];
-        v.extend_from_slice(args);
-        v
-    })
-    .context(id)
-    .unwrap();
+    for (extension, args2) in test_set() {
+        test_report(model, name, extension, [args, args2].concat()).context(id.clone()).unwrap();
+    }
 }
 
 // TODO:
@@ -104,4 +85,40 @@ fn instantiations() {
 fn cargo_config() {
     run("cargo_config", "cargo_config", &[]);
     run("cargo_config_toml", "cargo_config_toml", &[]);
+}
+
+// It seems rustup is not installed in the docker image provided by cross.
+#[cfg_attr(target_env = "musl", ignore)]
+#[test]
+fn merge() {
+    let model = "merge";
+    let output_dir = auxiliary::FIXTURES_PATH.join("coverage-reports").join(model);
+    fs::create_dir_all(&output_dir).unwrap();
+    for (extension, args) in test_set() {
+        // TODO: On windows, the order of the instantiations in the generated coverage report will be different.
+        #[cfg(windows)]
+        if extension == "txt" || extension == "full.json" {
+            continue;
+        }
+
+        let workspace_root = auxiliary::test_project(model, model).unwrap();
+        let output_path = &output_dir.join(model).with_extension(extension);
+        cargo_llvm_cov()
+            .args(["--color", "never", "--no-report", "--features", "a"])
+            .current_dir(workspace_root.path())
+            .assert_success();
+        cargo_llvm_cov()
+            .args(["--color", "never", "--no-report", "--features", "b"])
+            .current_dir(workspace_root.path())
+            .assert_success();
+        cargo_llvm_cov()
+            .args(["--color", "never", "--no-run", "--output-path"])
+            .arg(output_path)
+            .args(args)
+            .current_dir(workspace_root.path())
+            .assert_success();
+
+        auxiliary::normalize_output(output_path, args).unwrap();
+        auxiliary::assert_output(output_path).unwrap();
+    }
 }